refactor(fritz): move VPN classes to proper modules
- Move FritzVpnCoordinator from vpn_coordinator.py to coordinator.py
- Move FritzVpnSwitch from vpn_switch.py to switch.py
- Update imports in vpn_data.py and switch.py
- Delete obsolete vpn_coordinator.py and vpn_switch.py

This fixes pylint C7411 errors about class placement.
