+++ homeassistant/components/fritz/coordinator.py
@@ -10,6 +10,13 @@ import re
 from typing import Any, TypedDict, cast
 from xml.etree.ElementTree import ParseError
 
+from fritzboxvpn import (
+    API_KEY_ACTIVE,
+    API_KEY_CONNECTED,
+    API_KEY_NAME,
+    FritzBoxVPNSession,
+)
+from fritzboxvpn.const import PROTOCOL_HTTP, PROTOCOL_HTTPS
 from fritzconnection import FritzConnection
 from fritzconnection.core.exceptions import FritzActionError
 from fritzconnection.lib.fritzcall import FritzCall
@@ -23,10 +30,12 @@ from homeassistant.components.device_tracker import (
     DEFAULT_CONSIDER_HOME,
     DOMAIN as DEVICE_TRACKER_DOMAIN,
 )
-from homeassistant.config_entries import ConfigEntry
+from homeassistant.config_entries import ConfigEntry, ConfigEntryState
+from homeassistant.const import CONF_HOST, CONF_PASSWORD, CONF_SSL, CONF_USERNAME
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady, HomeAssistantError
 from homeassistant.helpers import device_registry as dr, entity_registry as er
+from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.device_registry import CONNECTION_NETWORK_MAC
 from homeassistant.helpers.dispatcher import async_dispatcher_send
 from homeassistant.helpers.typing import StateType
@@ -44,7 +53,14 @@ from .const import (
     DOMAIN,
     FRITZ_AUTH_EXCEPTIONS,
     FRITZ_EXCEPTIONS,
+    LOG_MSG_VPN_CONNECTIONS_REMOVED,
     SCAN_INTERVAL,
+    VPN_AUTH_INDICATORS,
+    VPN_RETRY_AFTER_SECONDS,
+    VPN_STATUS_CONNECTED,
+    VPN_STATUS_DISABLED,
+    VPN_STATUS_ENABLED,
+    VPN_STATUS_UNKNOWN,
     MeshRoles,
 )
 from .helpers import ha_is_stopping
@@ -938,3 +954,116 @@ class AvmWrapper(FritzBoxTools):
             "X_AVM-DE_WakeOnLANByMACAddress",
             NewMACAddress=mac_address,
         )
+
+
+def vpn_auth_failed(error: BaseException) -> bool:
+    """Return True when the error indicates invalid VPN credentials."""
+    message = str(error).lower()
+    return any(indicator in message for indicator in VPN_AUTH_INDICATORS)
+
+
+def vpn_web_ui_protocol(config: dict[str, Any]) -> str:
+    """Return the WireGuard web UI protocol for the config entry."""
+    return cast(
+        str,
+        PROTOCOL_HTTPS if config.get(CONF_SSL, DEFAULT_SSL) else PROTOCOL_HTTP,
+    )
+
+
+class FritzVpnCoordinator(DataUpdateCoordinator[dict[str, Any]]):
+    """Coordinator for WireGuard VPN connections on a FRITZ!Box."""
+
+    def __init__(
+        self,
+        hass: HomeAssistant,
+        config: dict[str, Any],
+        *,
+        entry_id: str,
+    ) -> None:
+        """Initialize the WireGuard VPN coordinator."""
+        super().__init__(
+            hass,
+            _LOGGER,
+            name=f"{DOMAIN}_vpn",
+            update_interval=timedelta(seconds=SCAN_INTERVAL),
+        )
+        self.entry_id = entry_id
+        self._reauth_scheduled = False
+        self.fritz_session = FritzBoxVPNSession(
+            async_get_clientsession(hass),
+            config[CONF_HOST],
+            config[CONF_USERNAME],
+            config[CONF_PASSWORD],
+            protocol=vpn_web_ui_protocol(config),
+        )
+
+    def get_vpn_status(self, connection_uid: str) -> str:
+        """Return the VPN status string for a connection."""
+        if not self.data or connection_uid not in self.data:
+            return VPN_STATUS_UNKNOWN
+        conn = self.data[connection_uid]
+        if not conn.get(API_KEY_ACTIVE, False):
+            return VPN_STATUS_DISABLED
+        if conn.get(API_KEY_CONNECTED, False):
+            return VPN_STATUS_CONNECTED
+        return VPN_STATUS_ENABLED
+
+    def _schedule_reauth(self) -> None:
+        if self._reauth_scheduled or not self.entry_id:
+            return
+        entry = self.hass.config_entries.async_get_entry(self.entry_id)
+        if entry is None or entry.state != ConfigEntryState.LOADED:
+            return
+        self._reauth_scheduled = True
+        entry.async_start_reauth(self.hass)
+
+    def _update_failed(
+        self, err: Exception, *, unexpected: bool = False
+    ) -> UpdateFailed:
+        if vpn_auth_failed(err):
+            self._schedule_reauth()
+        prefix = "Unexpected error" if unexpected else "Error"
+        if unexpected:
+            _LOGGER.exception("WireGuard VPN data update failed")
+        message = f"{prefix} fetching WireGuard VPN data: {err}"
+        if vpn_auth_failed(err):
+            return UpdateFailed(message)
+        return UpdateFailed(message, retry_after=VPN_RETRY_AFTER_SECONDS)
+
+    def _log_removed_connections(
+        self, previous_uids: set[str], connections: dict[str, Any]
+    ) -> None:
+        removed = previous_uids - connections.keys()
+        if not removed:
+            return
+        names = [self.data.get(uid, {}).get(API_KEY_NAME, uid) for uid in removed]
+        _LOGGER.warning(LOG_MSG_VPN_CONNECTIONS_REMOVED, names or list(removed))
+
+    async def _async_update_data(self) -> dict[str, Any]:
+        previous_uids = set(self.data) if self.data else set()
+        try:
+            connections = await self.fritz_session.async_get_vpn_connections()
+        except (ConnectionError, ValueError, TimeoutError) as err:
+            raise self._update_failed(err) from err
+        except Exception as err:
+            raise self._update_failed(err, unexpected=True) from err
+
+        self._log_removed_connections(previous_uids, connections)
+        self._reauth_scheduled = False
+        return cast(dict[str, Any], connections)
+
+    async def async_toggle_vpn(self, connection_uid: str, enable: bool) -> bool:
+        """Toggle a WireGuard VPN connection on or off."""
+        try:
+            return cast(
+                bool,
+                await self.fritz_session.async_toggle_vpn(connection_uid, enable),
+            )
+        except Exception as err:
+            if vpn_auth_failed(err):
+                self._schedule_reauth()
+            raise
+
+    async def async_close(self) -> None:
+        """Close the WireGuard VPN session."""
+        await self.fritz_session.async_close()
